app/requirements.txt
tika==2.6.0
pyexcel-ods==3 0.6.1
fastDamerauLevenshtein==1.0.7
pandas==1.5.3
networkx==3.0
python-sqlite-cache==0.0.2